Make createPersistentStore independent from store stucture.
