+++ src/App.tsx
@@ -1,22 +1,57 @@
 import classnames from 'classnames'
 import React from 'react'
+import _ from 'lodash/fp'
 import {SplashScreen} from './SplashScreen'
-import {BackendState} from './common/backend-state'
+import {
+  BackendState,
+  defaultBackendData,
+  migrationsForBackendData,
+  StoreBackendData,
+} from './common/backend-state'
 import {rendererLog} from './common/logger'
-import {createPersistentStore} from './common/store'
-import {WalletState} from './common/wallet-state'
+import {createPersistentStore, Store} from './common/store'
+import {WalletState, defaultWalletData, StoreWalletData} from './common/wallet-state'
 import {config} from './config/renderer'
 import {Router} from './layout/Router'
 import {Sidebar} from './layout/Sidebar'
 import {RouterState} from './router-state'
-import {SettingsState} from './settings-state'
-import {TokensState} from './tokens/tokens-state'
+import {SettingsState, defaultSettingsData, StoreSettingsData} from './settings-state'
+import {TokensState, defaultTokensData, StoreTokensData} from './tokens/tokens-state'
 import {TransactionHistoryService} from './wallets/history'
 import {createWeb3} from './web3'
 import './App.scss'
 
 const web3 = createWeb3(config.rpcAddress)
-const store = createPersistentStore()
+
+// TODO: move migrations and default data related code after restructuring store/state related code
+
+export type StoreData = StoreWalletData & StoreSettingsData & StoreTokensData & StoreBackendData
+
+const defaultData: StoreData = _.mergeAll([
+  defaultWalletData,
+  defaultSettingsData,
+  defaultTokensData,
+  defaultBackendData,
+])
+
+const mergeMigrations = _.mergeAllWith(
+  (
+    objValue: undefined | ((store: Store<StoreData>) => void),
+    srcValue: (store: Store<StoreData>) => void,
+  ) => {
+    if (objValue === undefined) {
+      return srcValue
+    } else {
+      return (store: Store<StoreData>): void => {
+        objValue.call(undefined, store)
+        srcValue.call(undefined, store)
+      }
+    }
+  },
+)
+
+const migrations = mergeMigrations([migrationsForBackendData])
+const store = createPersistentStore({defaults: defaultData, migrations})
 
 const AppContent: React.FC = () => {
   const backendState = BackendState.useContainer()
